+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Added support for `Series.dt.microsecond` and `Series.dt.nanosecond`.
 - Added support for `Index.is_unique` and `Index.has_duplicates`.
 - Added support for `Index.equals`.
+- Added support for `Index.value_counts`.
 
 #### Improvements
 - Removed the public preview warning message upon importing Snowpark pandas.